--- src/components/ProductList_Nav/Accordion.jsx
@@ -1,17 +1,0 @@
-export function Accordion() {
-  return (
-    <div>
-      <div>
-        {/* 필터 초기화  */}
-        <span>필터</span>
-        <button>
-          <img
-            src="../../assets/ProductListImage/resetButton.svg"
-            alt="초기화 버튼 이미지"
-          />
-          <span>초기화</span>
-        </button>
-      </div>
-    </div>
-  );
-}
