--- src/components/ProductList_Nav/Accordion_List.jsx
@@ -1,16 +1,0 @@
-export function Accordion_List({ list_name, count }) {
-  return (
-    <li>
-      <a className={`.willRouter`} href="/">
-        <button>
-          <img
-            alt="해당 리스트 체크하는 버튼"
-            src="../../assets/ProductListImage/Check_off.svg"
-          ></img>
-        </button>
-        <span>{list_name}</span>
-        <span>{count}</span>
-      </a>
-    </li>
-  );
-}
